Refaktorering av forskyvning barnehageplass
