+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vedtak/vedtaksperiode/Opphørsperiode.kt
@@ -14,8 +14,11 @@ import no.nav.familie.ks.sak.common.util.toYearMonth
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.domene.Vedtak
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode.domene.VedtaksperiodeMedBegrunnelser
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkår
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.VilkårRegelsett
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkårsvurdering
-import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.hentRegelsettBruktIVilkår
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.regelsett.lov2021.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.regelsett.lov2024.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør2024
 import no.nav.familie.ks.sak.kjerne.beregning.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.NasjonalEllerFellesBegrunnelse
 import no.nav.familie.ks.sak.kjerne.brev.begrunnelser.tilVedtaksbegrunnelse
@@ -86,10 +89,10 @@ fun slåSammenOpphørsperioder(alleOpphørsperioder: List<Opphørsperiode>): Lis
                 acc[acc.lastIndex] =
                     forrigeOpphørsperiode.copy(
                         periodeTom =
-                            maxOfOpphørsperiodeTom(
-                                forrigeOpphørsperiode.periodeTom,
-                                nesteOpphørsperiode.periodeTom,
-                            ),
+                        maxOfOpphørsperiodeTom(
+                            forrigeOpphørsperiode.periodeTom,
+                            nesteOpphørsperiode.periodeTom,
+                        ),
                     )
             }
 
@@ -118,9 +121,18 @@ private fun finnOpphørsperiodeEtterSisteUtbetalingsperiode(
 
     val erFramtidigOpphørPgaBarnehageplass =
         vilkårsvurdering.personResultater.any {
-            it.vilkårResultater.any {
-                it.vilkårType == Vilkår.BARNEHAGEPLASS && it.søkerHarMeldtFraOmBarnehageplass == true &&
-                    it.periodeTom?.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør() == sisteUtbetalingsperiodeTom
+            val vilkårResultaterPåPerson = it.vilkårResultater
+            val regelsett = vilkårResultaterPåPerson.hentRegelsettBruktIVilkår() ?: return@any false
+
+            vilkårResultaterPåPerson.any {
+                val periodeTom = it.periodeTom
+                val periodeTomForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør =
+                    when (regelsett) {
+                        VilkårRegelsett.LOV_AUGUST_2021 -> periodeTom?.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør()
+                        VilkårRegelsett.LOV_AUGUST_2024 -> periodeTom?.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør2024()
+                    }
+
+                it.vilkårType == Vilkår.BARNEHAGEPLASS && it.søkerHarMeldtFraOmBarnehageplass == true && periodeTomForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør == sisteUtbetalingsperiodeTom
             }
         }
