+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vilkårsvurdering/regelsett/lov2021/Graderingsforskjell.kt
@@ -1,0 +1,11 @@
+package no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.regelsett.lov2021
+
+enum class Graderingsforskjell {
+    IngenUtbetalingGrunnetFullBarnehageplassTilØking,
+    Øking,
+    ØkingGrunnetSluttIBarnehage,
+    ReduksjonGårTilIngenUtbetaling,
+    Reduksjon,
+    Lik,
+    IngenUtbetalingGrunnetFørsteperiodeTilØking,
+}
