+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vilkårsvurdering/regelsett/lov2024/ForskyvBarnehageplass2024.kt
@@ -1,0 +1,108 @@
+package no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.regelsett.lov2024
+
+import no.nav.familie.ks.sak.common.tidslinje.Periode
+import no.nav.familie.ks.sak.common.tidslinje.tilTidslinje
+import no.nav.familie.ks.sak.common.tidslinje.utvidelser.kombiner
+import no.nav.familie.ks.sak.common.tidslinje.utvidelser.tilPerioder
+import no.nav.familie.ks.sak.common.tidslinje.utvidelser.tilPerioderIkkeNull
+import no.nav.familie.ks.sak.common.util.*
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.VilkårResultat
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.tilTidslinje
+import no.nav.familie.ks.sak.kjerne.beregning.domene.hentProsentForAntallTimer
+import java.math.BigDecimal
+import java.time.LocalDate
+
+fun List<VilkårResultat>.forskyvBarnehageplassVilkår2024(): List<Periode<VilkårResultat>> {
+    return tilBarnehageplassVilkårMedGraderingsforskjellMellomPerioder2024()
+        .map {
+            Periode(
+                verdi = it.vilkårResultat,
+                fom =
+                    it.vilkårResultat.periodeFom
+                        .tilForskøvetFomBasertPåGraderingsforskjell2024(it.graderingsforskjellMellomDenneOgForrigePeriode),
+                tom =
+                    it.vilkårResultat.periodeTom
+                        .tilForskøvetTomBasertPåGraderingsforskjell2024(it.graderingsforskjellMellomDenneOgNestePeriode),
+            )
+        }.filter { (it.fom ?: TIDENES_MORGEN).isBefore(it.tom ?: TIDENES_ENDE) }
+        .filtrerBortOverlappendePerioderMedMaksGradering()
+}
+
+fun LocalDate?.tilForskjøvetTomMånedForSisteUtbetalingsperiodePgaFremtidigOpphør2024() = this?.tilForskøvetTomBasertPåGraderingsforskjell2024(Graderingsforskjell.Reduksjon)?.toYearMonth()
+
+private fun List<VilkårResultat>.tilBarnehageplassVilkårMedGraderingsforskjellMellomPerioder2024(): List<BarnehageplassVilkårMedGraderingsforskjellMellomPerioder<VilkårResultat>> {
+    val vilkårResultatListeMedNullverdierForHullITidslinje: List<VilkårResultat?> =
+        this
+            .tilTidslinje()
+            .tilPerioder()
+            .map { it.verdi }
+
+    return vilkårResultatListeMedNullverdierForHullITidslinje
+        .fold(emptyList()) { acc: List<BarnehageplassVilkårMedGraderingsforskjellMellomPerioder<VilkårResultat?>>, vilkårResultat ->
+            val vilkårResultatIForrigePeriode = acc.lastOrNull()
+
+            val graderingsforskjellMellomDenneOgForrigePeriode =
+                vilkårResultat.hentGraderingsforskjellMellomDenneOgForrigePeriode2024(vilkårResultatIForrigePeriode)
+
+            val accMedForrigeOppdatert =
+                if (vilkårResultatIForrigePeriode == null) {
+                    acc
+                } else {
+                    acc.dropLast(1) +
+                        vilkårResultatIForrigePeriode
+                            .copy(graderingsforskjellMellomDenneOgNestePeriode = graderingsforskjellMellomDenneOgForrigePeriode)
+                }
+
+            accMedForrigeOppdatert +
+                BarnehageplassVilkårMedGraderingsforskjellMellomPerioder(
+                    vilkårResultat = vilkårResultat,
+                    graderingsforskjellMellomDenneOgForrigePeriode = graderingsforskjellMellomDenneOgForrigePeriode,
+                    graderingsforskjellMellomDenneOgNestePeriode = Graderingsforskjell.Reduksjon,
+                )
+        }.filtrerBortNullverdier()
+}
+
+private fun VilkårResultat?.hentGraderingsforskjellMellomDenneOgForrigePeriode2024(
+    vilkårResultatForrigePeriode: BarnehageplassVilkårMedGraderingsforskjellMellomPerioder<VilkårResultat?>?,
+): Graderingsforskjell {
+    val graderingForrigePeriode =
+        vilkårResultatForrigePeriode?.vilkårResultat?.let { hentProsentForAntallTimer(vilkårResultatForrigePeriode.vilkårResultat.antallTimer) }
+            ?: BigDecimal.ZERO
+    val graderingDennePerioden = this?.let { hentProsentForAntallTimer(this.antallTimer) } ?: BigDecimal.ZERO
+    return when {
+        graderingForrigePeriode > graderingDennePerioden -> Graderingsforskjell.Reduksjon
+        graderingForrigePeriode < graderingDennePerioden -> Graderingsforskjell.Øking
+        else -> Graderingsforskjell.Lik
+    }
+}
+
+
+private fun List<Periode<VilkårResultat>>.filtrerBortOverlappendePerioderMedMaksGradering() =
+    map { listOf(it).tilTidslinje() }
+        .kombiner { vilkårResultater -> vilkårResultater.minByOrNull { it.antallTimer ?: BigDecimal.ZERO } }
+        .tilPerioderIkkeNull()
+
+private fun LocalDate?.tilForskøvetTomBasertPåGraderingsforskjell2024(graderingsforskjellMellomDenneOgNestePeriode: Graderingsforskjell) =
+    this?.let { tomDato ->
+        when (graderingsforskjellMellomDenneOgNestePeriode) {
+            Graderingsforskjell.Lik,
+            -> tomDato.minusMonths(1).sisteDagIMåned()
+
+            Graderingsforskjell.Øking,
+            -> tomDato.sisteDagIMåned()
+
+            Graderingsforskjell.Reduksjon -> tomDato.plusDays(1).sisteDagIMåned()
+        }
+    }
+
+private fun LocalDate?.tilForskøvetFomBasertPåGraderingsforskjell2024(graderingsforskjellMellomDenneOgForrigePeriode: Graderingsforskjell) =
+    this?.let { fomDato ->
+        when (graderingsforskjellMellomDenneOgForrigePeriode) {
+            Graderingsforskjell.Lik,
+            Graderingsforskjell.Øking,
+            -> fomDato.førsteDagIInneværendeMåned()
+
+            Graderingsforskjell.Reduksjon,
+            -> fomDato.plusMonths(1).førsteDagIInneværendeMåned()
+        }
+    }
