+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vilkårsvurdering/regelsett/lov2024/Graderingsforskjell.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.regelsett.lov2024
+
+enum class Graderingsforskjell {
+    Øking,
+    Reduksjon,
+    Lik,
+}
